+++ NEWS.md
@@ -252,8 +252,9 @@ Then, we fixed this by using the canonical URL of the CRAN package "https://cran
 - We implemented a new function `EM3.cov`, which solves a multi-kernel linear mixed-effects model considering covariance structure between random effects. We added the function to `NAMESPACE` file.
 
 
-# June 13, 2025, RAINBOWR version 0.1.39
+# June 26, 2025, RAINBOWR version 0.1.39
 ## Major changes (new functions, etc...)
 - We fixed the error in `calcGRM` function with `methodGRM == "linear` option.
 - We fixed the error in `estPhylo` and `plotPhyloTree` functions when drawing phylogenetic trees.
 - We fixed the error in `estNetwork` and `plotHaploNetwork` functions when drawing haplotype network.
+- We added new options for `EM3.cov` so that the user can determine initial weights for the genetic variance from outside.